updated clean.sh to delete .postgres file properly

+++ scripts/flake/clean.sh
@@ -5,8 +5,7 @@ targets=(
   ".postgres"
   ".venv"
   "build"
-  "ctgov-compliance-web.egg-info"
-  ".aws-profile"
+  "ctgov_compliance_web.egg-info"
   ".env"
   "blazegraph.jar"
   ".blazegraph.pid"
@@ -16,14 +15,23 @@ targets=(
   "rules.log"
 )
 
-# Loop through each target and delete it if it exists
-for target in "${targets[@]}"; do
+# Function to safely delete a target
+delete_target() {
+  local target=$1
   if [ -e "$target" ]; then
     echo "Deleting: $target"
-    rm -rf "$target"
+    if ! rm -rf "$target" 2>/dev/null; then
+      echo "Permission denied, trying with sudo..."
+      rm -rf "$target"
+    fi
   else
     echo "Not found: $target"
   fi
+}
+
+# Loop through each target and delete it
+for target in "${targets[@]}"; do
+  delete_target "$target"
 done
 
 # Find and kill Postgres listening on port 5464
@@ -33,6 +41,29 @@ if [ -n "$PIDS" ]; then
   echo "Killing Postgres processes on port 5464 with PID $PIDS: "
   echo "$PROCESS"
   kill $PIDS
+  for pid in $PIDS; do
+    while kill -0 $pid 2>/dev/null; do
+      echo "Waiting for Postgres process $pid to terminate..."
+      sleep 1
+    done
+  done
 else
   echo "No Postgres processes found on port 5464"
+fi
+
+# Find and kill Java process listening on port 9999
+JAVA_PIDS=$(lsof -t -i :9999 -sTCP:LISTEN -a -c java)
+JAVA_PROCESS=$(lsof -i :9999 -sTCP:LISTEN -a -c java | head -n 2)
+if [ -n "$JAVA_PIDS" ]; then
+  echo "Killing Java processes on port 9999 with PID $JAVA_PIDS:"
+  echo -e "Full entry info:\n$JAVA_PROCESS"
+  kill $JAVA_PIDS
+  for pid in $JAVA_PIDS; do
+    while kill -0 $pid 2>/dev/null; do
+      echo "Waiting for Java process $pid to terminate..."
+      sleep 1
+    done
+  done
+else
+  echo "No Java processes found on port 9999"
 fi
